Remove use of optparse, simply use sys.argv for runtests.py
OptionParser seems better designed for when we want explicit set of command
line arguments to pass. In this case, we simply want to pass through
any command line args straight to nose runner, allowing it to
handle them as necessary.

Prior to this, it'd be impossible to pass nose args via the command line.